Migrate Transitions to styled-component (#344)
