+++ src/index.tsx
@@ -41,7 +41,7 @@ export { default as TimePicker } from 'timePicker'
 export { default as Title } from 'title'
 export { default as ToggleButton, ToggleButtonStatus } from 'toggleButton'
 export { default as TopBar } from 'topBar'
-export { default as Transitions } from 'transitions'
+export { default as Transitions, AnimationType } from 'transitions'
 export { default as TripCard } from 'tripCard'
 export { default as UneditableTextField } from 'uneditableTextField'
 export { default as Why } from 'why'
